refactor: 패키지명 ui -> view로 변경
